Added buffs

+++ Assets/Scripts/Effect.cs
@@ -12,11 +12,11 @@ namespace Assets.Scripts
         public Constants.EffectType Type { get; set; }
         public Constants.BuffTypes BuffType { get; set; }
         public Constants.StatusTypes StatusType { get; set; }
-        public int Power { get; set; }
+        public float Power { get; set; }
         public int Duration { get; set; }
 
         public Effect(Constants.EffectType type, Constants.BuffTypes buffType = Constants.BuffTypes.None,
-            Constants.StatusTypes statusType = Constants.StatusTypes.None, int power = 0, int duration = 2)
+            Constants.StatusTypes statusType = Constants.StatusTypes.None, float power = 0, int duration = 2)
         {
             Type = type;
             BuffType = buffType;
